feat: Move feature to existing target goal feature
