--- src/extension/features/budget/goal-display-filter/index.css
@@ -1,3 +1,0 @@
-.tk-goal-sum-amount {
-  margin-left: auto;
-}
