--- src/extension/features/budget/goal-display-filter/settings.js
@@ -1,8 +1,0 @@
-module.exports = {
-  name: 'SumGoalDisplay',
-  type: 'checkbox',
-  default: false,
-  section: 'budget',
-  title: 'Add Goal Sum To Category Group',
-  description: 'Add a sum to each category group for goal amounts.',
-};
